Add initial transitions to show/hide Image Slider
These don't work exactly as expected, they're smooth enough; but the slide down
transition on the list doesn't work when showing the Image Slider. This might be
due to fading in the image when loaded. Because the image may be faded in _after_
the list slides down the slide down effect is lost. As a potential work around,
the list can be slide down _after_ the selected image is faded in.